Move react-form tests
