+++ packages/react-form/tests/useField.test.tsx
@@ -2,9 +2,9 @@ import * as React from 'react'
 import { describe, expect, it, vi } from 'vitest'
 import { render, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { useForm } from '../index'
+import { useForm } from '../src/index'
 import { sleep } from './utils'
-import type { FieldApi, FormApi } from '../index'
+import type { FieldApi, FormApi } from '../src/index'
 
 const user = userEvent.setup()
