+++ packages/react-form/tests/useForm.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { describe, expect, it, vi } from 'vitest'
 import { render, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { useForm } from '../index'
+import { useForm } from '../src/index'
 import { sleep } from './utils'
 
 const user = userEvent.setup()
