--- packages/react-form/tsconfig.50.json
@@ -1,4 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "exclude": ["src/tests/**"]
-}
