+++ packages/react-form/tsconfig.docs.json
@@ -1,9 +1,9 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["src/tests/**/*"],
   "compilerOptions": {
     "paths": {
       "@tanstack/form-core": ["../form-core/src"]
     }
-  }
+  },
+  "exclude": ["tests"]
 }
