+++ packages/react-form/tsconfig.json
@@ -7,11 +7,5 @@
       "@tanstack/form-core": ["../form-core/src"]
     }
   },
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.tsx",
-    ".eslintrc.cjs",
-    "test-setup.ts",
-    "vite.config.ts"
-  ]
+  "include": ["src", "tests", ".eslintrc.cjs", "vite.config.ts"]
 }
