+++ packages/react-form/tsconfig.legacy.json
@@ -7,6 +7,5 @@
       "@tanstack/form-core": ["../form-core/src"]
     }
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["src/tests/**"]
+  "include": ["src"]
 }
