+++ packages/react-form/vite.config.ts
@@ -6,10 +6,10 @@ const config = defineConfig({
   plugins: [react()],
   test: {
     name: 'react-form',
-    dir: './src',
+    dir: './tests',
     watch: false,
     environment: 'jsdom',
-    setupFiles: ['test-setup.ts'],
+    setupFiles: ['./tests/test-setup.ts'],
     coverage: { enabled: true, provider: 'istanbul', include: ['src/**/*'] },
     typecheck: { enabled: true },
   },
